requirement.txt
absl-py==1.4.0
airsim==1.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.2
bleach==6.0.0
brotlipy==0.7.0
cachetools==5.3.0
cairocffi==1.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.1.0
cryptography==39.0.1
cycler==0.11.0
debugpy==1.6.6
decorator==4.4.2
deeprobust==0
defusedxml==0.7.1
dgl==0.9.1
easydict==1.10
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
ffmpeg==1.4
flit_core==3.8.0
fonttools==4.34.4
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.3.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.1
gym==0.11.0
gym-flock==0.0.1
gym-notices==0.0.7
hashids==1.3.1
idna==3.4
igraph==0.10.3
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.14.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter_core==5.2.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
karateclub==1.3.3
kiwisolver==1.4.4
lazy_loader==0.2
Levenshtein==0.20.9
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.4.0
matplotlib-inline==0.1.6
minepy==1.2.6
miniful==0.0.6
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.4.12
numba==0.56.4
numpy==1.22.4
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyFUME==0.2.25
pyglet==1.5.26
Pygments==2.14.0
PyGSP==0.5.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-Levenshtein==0.20.9
python-louvain==0.16
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.3.1
QtPy==2.3.0
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
SciencePlots==2.0.1
scikit-image==0.20.0
scikit-learn==1.1.1
scipy==1.9.0
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==67.4.0
simpful==2.9.0
sip==6.7.7
six==1.16.0
smart-open==6.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
terminado==0.15.0
texttable==1.6.7
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==1.12.0
torch-cluster==1.6.0
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.14
torch-spline-conv==1.2.1
torchaudio==0.12.0
torchsummaryX==1.3.0
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
wheel==0.38.4
widgetsnbextension==3.6.1
zipp==3.14.0
